+++ src/Builders/PHPFileBuilder.php
@@ -12,9 +12,9 @@ use RonasIT\Larabuilder\NodeTraverser;
 use RonasIT\Larabuilder\Printer;
 use RonasIT\Larabuilder\ValueOptions\MethodParams;
 use RonasIT\Larabuilder\Visitors\AddImports;
-use RonasIT\Larabuilder\Visitors\AddMethod;
 use RonasIT\Larabuilder\Visitors\AddTraits;
-use RonasIT\Larabuilder\Visitors\InsertCodeToMethod;
+use RonasIT\Larabuilder\Visitors\MethodVisitors\AddMethod;
+use RonasIT\Larabuilder\Visitors\MethodVisitors\InsertCodeToMethod;
 use RonasIT\Larabuilder\Visitors\PropertyVisitors\AddArrayPropertyItem;
 use RonasIT\Larabuilder\Visitors\PropertyVisitors\RemoveArrayPropertyItem;
 use RonasIT\Larabuilder\Visitors\PropertyVisitors\SetProperty;
