+++ src/Visitors/MethodVisitors/AbstractMethodVisitor.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace RonasIT\Larabuilder\Visitors\MethodVisitors;
+
+use PhpParser\Node\Stmt\Class_;
+use PhpParser\Node\Stmt\Enum_;
+use PhpParser\Node\Stmt\Trait_;
+use RonasIT\Larabuilder\Visitors\BaseNodeVisitorAbstract;
+
+abstract class AbstractMethodVisitor extends BaseNodeVisitorAbstract
+{
+    protected array $allowedParentNodesTypes = [
+        Class_::class,
+        Trait_::class,
+        Enum_::class,
+    ];
+}
